# go-dbu-api-develop/docs/swagger.yaml
basePath: /
definitions:
...
  models.Links:
    properties:
      next:
        type: string
      prev:
        type: string
    type: object
...
  models.Metadata:
    properties:
      limit:
        type: integer
      links:
        $ref: '#/definitions/models.Links'
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      metadata:
        $ref: '#/definitions/models.Metadata'
      msg:
        type: string
      type:
        type: string
    type: object
...
  room_assignment.RequestRoomAssignment:
    properties:
      residence_id:
        type: string
    type: object
  room_assignment.ResponseAutomaticAssignationResidence:
    properties:
      register_id:
        type: string
    type: object
...
info:
  contact:
    email: juanm.campos@unas.edu.pe
    name: API Support
  description: DBU API
  license:
    name: Comercial
    url: http://bjungle.net/licenses
  termsOfService: https://dbu.com/terms/
  title: DBU API
  version: "1.0"
paths:
...
  /api/v1/area_medica/reporte/consultas:
    get:
      consumes:
      - application/json
      description: Método que permite obtener el reporte de consultas de enfermería
      parameters:
      - description: Área médica (obligatoria)
        in: query
        name: area_medica
        required: true
        type: string
      - description: 'Fecha de inicio (opcional) - formato: YYYY-MM-DD'
        in: query
        name: fecha_inicio
        type: string
      - description: 'Fecha de fin (opcional) - formato: YYYY-MM-DD'
        in: query
        name: fecha_fin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
      summary: Obtener reporte de consultas de enfermería
      tags:
      - Reportes
...
  /v1/automatizacion/asignacion-cuartos:
    post:
      consumes:
      - application/json
      description: Endpoint para ejecutar la asignación automática de cuartos en una
        residencia específica
      parameters:
      - description: ID de la residencia para asignación
        in: body
        name: RequestRoomAssignment
        required: true
        schema:
          $ref: '#/definitions/room_assignment.RequestRoomAssignment'
      produces:
      - application/json
      responses:
        "200":
          description: Asignación exitosa
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/room_assignment.ResponseAutomaticAssignationResidence'
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "400":
          description: Error en la solicitud
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "401":
          description: Error de autenticación
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                code:
                  type: integer
                data: {}
                error:
                  type: boolean
                msg:
                  type: string
                type:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Asignación automática de cuartos en residencias
      tags:
      - Automatización
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"